Redoing the previous attempt with latest code. (#558)
[ios_acls] Fix AttributeError on dictionary check

This is a redo of #516 as I messed up the repo while updating it to the latest code base.
Check if proto_options is actually a dictionary instead of simply checking if it exists.
SUMMARY

This fixes the error AttributeError: 'bool' object has no attribute 'keys', which is a result of simply checking if proto_options exists, rather than making sure it is actually a dictionary.
Adding protocol_options to every entry isn't an option as that results in Unsupported attribute for standard ACL - protocol_options.
ISSUE TYPE


Bugfix Pull Request

COMPONENT NAME

ios_acls
ADDITIONAL INFORMATION


An example task that results in the error would be
- name: CISCO | IOS | Deploying ACL template
  cisco.ios.ios_acls:
    state: overridden
    config:
      - afi: ipv4
        acls:
          - name: TEST_ACL
            acl_type: extended
            aces:
              - sequence: 10
                grant: permit
                protocol_options:
                  udp: 'yes'
                source:
                  any: 'yes'
                  port_protocol:
                    eq: 68
                destination:
                  any: 'yes'
                  port_protocol:
                    eq: 69

Reviewed-by: Sagar Paul <sagpaul@redhat.com>